Add Commission model
ran `rails generate model commissions commission_percent:float
user:references`

ran `rails db:migrate`

added relation to User

The commission model will be used to store special comissions for
customers not on the normal rate.

Note: for some reason there is a bug that caused the has_one relation to
not find the correct model name for some reason, this was fixed by
adding the class_name argument. This is noted in the source code.