Add purchaseTransaction model
ran `rails generate model PurchaseTransaction purchase:references
inventoryTransaction:references`

ran `rails db:migrate`

added relations to purchase and inventoryTransaction

The purchaseTransaction model is needed as a join table so that
purchases can be linked to changes in inventory.

The inventoryTransaction model needed to have the foreign_key set on the
has_many relation with purchaseTransaction model as the
purchaseTransaction migration created the column named
'inventoryTransaction_id' but the has_many relation looked for a column
named 'inventory_transaction_id' seems to be a odd bug or someting.